# static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Notes App API",
    "version": "1.0.0",
    "description": "This API allows users to register, login, and perform CRUD operations on their notes."
  },
  "servers": [
    {
      "url": "https://notes-app-3-8hk7.onrender.com",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Note": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "title": { "type": "string" },
          "content": { "type": "string" },
          "category": { "type": "string" },
          "pinned": { "type": "boolean" }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "username": { "type": "string" },
          "password": { "type": "string" }
        }
      }
    }
  },
  "paths": {
    "/register": {
      "post": {
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/User" }
            }
          }
        },
        "responses": {
          "200": { "description": "User registered successfully" },
          "400": { "description": "Username already exists" }
        }
      }
    },
    "/login": {
      "post": {
        "summary": "Login an existing user and get JWT token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/User" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "JWT token returned",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/notes": {
      "get": {
        "summary": "Get all notes of logged-in user",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of notes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Note" }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      },
      "post": {
        "summary": "Create a new note",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Note" }
            }
          }
        },
        "responses": {
          "200": { "description": "Note created successfully" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/notes/{id}": {
      "put": {
        "summary": "Update a note",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Note" }
            }
          }
        },
        "responses": {
          "200": { "description": "Note updated successfully" },
          "401": { "description": "Unauthorized" }
        }
      },
      "delete": {
        "summary": "Delete a note",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "Note deleted successfully" },
          "401": { "description": "Unauthorized" }
        }
      }
    }
  }
}